chore: theme-provider -> theme; cleanup and story updates for colors
