+++ packages/ui/src/theme/index.ts
@@ -1,1 +1,2 @@
 export * from './ThemeProvider'
+export * from './defaults'
